requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
pydantic==2.10.4
python-multipart==0.0.20

passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4